% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempConversion.R
\name{C_to_F}
\alias{C_to_F}
\title{Celsius conversion}
\usage{
C_to_F(c_temp)
}
\arguments{
\item{c_temp}{The temperature in degrees Celsius}
}
\value{
The temperature in degrees Fahrenheit
}
\description{
Convert degrees Celsius temperatures to degrees Fahrenheit
}
\examples{
temp1 <- C_to_F(22);
temp2 <- C_to_F(c(-2,12,23));
}
